requirement.txt
accelerate
psycopg2
entence-transformers
faiss
chromadb